src/image_processor/requirements.txt
boto3==1.37.12
Pillow==11.1.0
